Refactor ArticleCard component to ArticleItem and update layout and image handling
